+++ src/main/java/de/uros/citlab/module/workflow/RunTrainingSemiSupervised.java
@@ -377,7 +377,7 @@ public class RunTrainingSemiSupervised extends ParamTreeOrganizer {
                 try {
                     FileUtils.deleteQuietly(folderSnipets);
                     folderSnipets.mkdirs();
-                    int res = TrainDataUtil.runCreateTraindata(folderPageXml, folderSnipets, charMap, props);
+                    int res = TrainDataUtil.runCreateTraindata(folderPageXml, folderSnipets, charMap, props,null);
                     if (res == 0) {
                         LOG.warn("copy folder - assume that folder " + folderPageXml.getAbsolutePath() + " contains snipets.");
                         FileUtils.copyDirectory(folderPageXml, folderSnipets);