+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+# 2.2.2
+* bugfix training: shuffle training list so that order is "randomly"
+* bugfix surrogates: if a line contains surrogates, it is ignored in training and validation
+* bugfix empty validation: If a validation set is given, but no lines are in, throw exception
+* feature notification: errors while creating training/validation data are reported via notification
+
+# 2.2.1
+* bugfix LA Advanced: when region is given and rotated text lines are found, lines will now not rotated out of image 
+
+# 2.2.0
+* update to newer versions for bugfix in planets sources
+
 # 2.1.1
 * bugfix training of HTR+ with base model